Use full path of the resource in groupId
Previously, only the parent directory and the library name
were taken into account when generating name for the dependency.

This meant when two resources were included, with following paths:

one/lib/dependency.jar
other/lib/dependency.jar

then only one was included as their idientfiers were the same.

Instead of only using the name of the parent folder, use the
absolute path of the library to generate unique ids for all
the resources.